add absolute path finish googleAuth UI
