+++ .idea/workspace.xml
@@ -5,14 +5,16 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="ec802122-7679-4b91-b9e1-dac82f02c137" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/src/assets/googleIcon.png" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/components/GoogleAuth/GoogleAuth.tsx" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/.env" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/components/GoogleAuth/GoogleAuthStyle.ts" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/components/GoogleAuth/colors.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.eslintrc.js" beforeDir="false" afterPath="$PROJECT_DIR$/.eslintrc.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/package-lock.json" beforeDir="false" afterPath="$PROJECT_DIR$/package-lock.json" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/package.json" beforeDir="false" afterPath="$PROJECT_DIR$/package.json" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/App.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/App.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/components/Sample/Sample.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/Sample/Sample.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/GoogleAuth/GoogleAuth.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/GoogleAuth/GoogleAuth.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/localization/en/en.json" beforeDir="false" afterPath="$PROJECT_DIR$/src/localization/en/en.json" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/redux/store.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/redux/store.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/tsconfig.json" beforeDir="false" afterPath="$PROJECT_DIR$/tsconfig.json" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -57,8 +59,9 @@
 }</component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
-      <recent name="C:\IT\intership\Devs_Heads_fe\src\assets" />
+      <recent name="C:\IT\intership\Devs_Heads_fe\src\components\GoogleAuth" />
       <recent name="C:\IT\intership\Devs_Heads_fe\src" />
+      <recent name="C:\IT\intership\Devs_Heads_fe\src\assets" />
       <recent name="C:\IT\intership\Devs_Heads_fe\src\components\hoc" />
       <recent name="C:\IT\intership\Devs_Heads_fe\src\components" />
     </key>
@@ -81,7 +84,9 @@
       <workItem from="1660222398343" duration="4692000" />
       <workItem from="1660231856866" duration="9313000" />
       <workItem from="1660288047554" duration="3767000" />
-      <workItem from="1660319338743" duration="608000" />
+      <workItem from="1660319338743" duration="8546000" />
+      <workItem from="1660345232541" duration="613000" />
+      <workItem from="1660414222933" duration="7718000" />
     </task>
     <servers />
   </component>
