+++ src/components/GoogleAuth/colors.ts
@@ -1,0 +1,3 @@
+export const BLUE = '#1E90FFFF';
+export const LIGHT_BLUE = '#45a2ff';
+export const WHITE = '#f5f5f5';
