+++ src/redux/store.ts
@@ -1,13 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
 import sampleReducer from './reducers/sampleSlice';
-import { pokemonApi } from '../components/GoogleAuth/GoogleAuth';
 
 export const store = configureStore({
 	reducer: {
-		[pokemonApi.reducerPath]: pokemonApi.reducer,
 		sample: sampleReducer,
 	},
-	middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(pokemonApi.middleware),
 });
 
 export type RootState = ReturnType<typeof store.getState>;
